chore(caseService): restructure CaseServiceTest into multiple tests
